requirements.txt
django-froala-editor
imagekit
slugify
asgiref==3.2.3
Django==3.0.1
django-appconf==1.0.3
django-ckeditor==5.8.0
django-imagekit==4.0.2
django-js-asset==1.2.2
pilkit==2.0
Pillow==7.0.0
python-slugify==4.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
